refactored visualizer code into visualizer classes
