+++ src/uvi/visualizations/FrameNetVisualizer.py
@@ -1,125 +1,27 @@
 """
-FrameNet Visualizer Base Class.
+FrameNet Visualizer Class.
 
-This module contains the base FrameNetVisualizer class that provides common functionality
-for creating different types of FrameNet semantic graph visualizations.
+This module contains the FrameNetVisualizer class that provides FrameNet-specific
+functionality for creating semantic graph visualizations.
 """
 
-from collections import defaultdict
-from pathlib import Path
-import networkx as nx
-import matplotlib.pyplot as plt
-
-# Optional Plotly import for enhanced interactivity
-try:
-    import plotly.graph_objects as go
-    PLOTLY_AVAILABLE = True
-except ImportError:
-    PLOTLY_AVAILABLE = False
+from .Visualizer import Visualizer
 
 
-class FrameNetVisualizer:
-    """Base class for FrameNet semantic graph visualizations."""
+class FrameNetVisualizer(Visualizer):
+    """FrameNet-specific visualizer with specialized coloring and information display."""
     
     def __init__(self, G, hierarchy, title="FrameNet Frame Hierarchy"):
-        """
-        Initialize the visualizer.
-        
-        Args:
-            G: NetworkX DiGraph
-            hierarchy: Frame hierarchy data
-            title: Title for visualizations
-        """
-        self.G = G
-        self.hierarchy = hierarchy
-        self.title = title
-        
-    def create_dag_layout(self):
-        """Create spring-based DAG layout for the graph."""
-        # Use NetworkX spring layout as base, but with DAG-aware enhancements
-        pos = nx.spring_layout(self.G, k=2.5, iterations=100, seed=42)
-        
-        # Apply vertical bias based on topological ordering for DAG structure
-        try:
-            topo_order = list(nx.topological_sort(self.G))
-            topo_positions = {node: i for i, node in enumerate(topo_order)}
-            
-            # Adjust Y coordinates to respect topological ordering while keeping spring positions
-            max_topo = len(topo_order) - 1
-            for node in pos:
-                if node in topo_positions:
-                    # Blend spring layout with topological ordering
-                    spring_y = pos[node][1]
-                    topo_y = 1.0 - (2.0 * topo_positions[node] / max_topo)  # Range from 1 to -1
-                    
-                    # Weight: 60% topological order, 40% spring layout
-                    blended_y = 0.6 * topo_y + 0.4 * spring_y
-                    pos[node] = (pos[node][0], blended_y)
-        
-        except nx.NetworkXError:
-            # If not a DAG (shouldn't happen), use pure spring layout
-            pass
-        
-        # Apply some spacing adjustments to avoid overlaps
-        self._adjust_positions_for_clarity(pos)
-        
-        return pos
-    
-    def create_taxonomic_layout(self):
-        """Create hierarchical layout based on depth levels."""
-        # Group nodes by depth levels for hierarchical layout
-        depth_nodes = defaultdict(list)
-        for node, data in self.G.nodes(data=True):
-            depth = data.get('depth', 0)
-            depth_nodes[depth].append(node)
-        
-        # Create hierarchical positions
-        pos = {}
-        for depth, nodes in depth_nodes.items():
-            n_nodes = len(nodes)
-            if n_nodes == 1:
-                x_positions = [0]
-            else:
-                # Spread nodes horizontally
-                spread = min(8, n_nodes * 1.5)
-                x_positions = [(i - (n_nodes-1)/2) * spread / n_nodes for i in range(n_nodes)]
-            
-            # Y position based on depth (negative to put roots at top)
-            y = -(depth * 3)
-            
-            for i, node in enumerate(sorted(nodes)):
-                pos[node] = (x_positions[i], y)
-        
-        return pos
-    
-    def _adjust_positions_for_clarity(self, pos):
-        """Adjust positions to improve clarity and reduce overlaps."""
-        nodes = list(pos.keys())
-        min_distance = 0.3  # Minimum distance between nodes
-        
-        # Simple separation adjustment
-        for i, node1 in enumerate(nodes):
-            for j, node2 in enumerate(nodes[i+1:], i+1):
-                x1, y1 = pos[node1]
-                x2, y2 = pos[node2]
-                
-                distance = ((x2 - x1) ** 2 + (y2 - y1) ** 2) ** 0.5
-                if distance < min_distance and distance > 0:
-                    # Push nodes apart
-                    dx = (x2 - x1) / distance
-                    dy = (y2 - y1) / distance
-                    
-                    adjustment = (min_distance - distance) / 2
-                    pos[node1] = (x1 - dx * adjustment, y1 - dy * adjustment)
-                    pos[node2] = (x2 + dx * adjustment, y2 + dy * adjustment)
+        """Initialize the FrameNet visualizer."""
+        super().__init__(G, hierarchy, title)
     
     def get_dag_node_color(self, node):
-        """Get color for a node based on DAG properties and node type."""
+        """Get color for a node based on DAG properties and FrameNet node type."""
         # Check if node has type information
         node_data = self.G.nodes.get(node, {})
         node_type = node_data.get('node_type', 'frame')
         
-        # Different colors for different node types
+        # Different colors for different FrameNet node types
         if node_type == 'lexical_unit':
             return 'lightyellow'  # Lexical units get yellow color
         elif node_type == 'frame_element':
@@ -138,20 +40,8 @@ class FrameNetVisualizer:
         else:
             return 'lightgray'    # Isolated nodes
     
-    def get_taxonomic_node_color(self, node):
-        """Get color for a node based on taxonomic depth."""
-        depth = self.G.nodes[node].get('depth', 0)
-        if depth == 0:
-            return 'lightblue'    # Root frames
-        elif depth == 1:
-            return 'lightgreen'  # Level 1 frames
-        elif depth == 2:
-            return 'lightyellow' # Level 2 frames
-        else:
-            return 'lightcoral'  # Deeper levels
-    
     def get_node_info(self, node):
-        """Get detailed information about a node."""
+        """Get detailed information about a FrameNet node."""
         if node not in self.hierarchy:
             return f"Node: {node}\nNo additional information available."
         
@@ -159,7 +49,7 @@ class FrameNetVisualizer:
         frame_info = data.get('frame_info', {})
         node_type = frame_info.get('node_type', 'frame')
         
-        # Different display format for different node types
+        # Different display format for different FrameNet node types
         if node_type == 'lexical_unit':
             info = [f"Lexical Unit: {frame_info.get('name', node)}"]
             info.append(f"Frame: {frame_info.get('frame', 'Unknown')}")
@@ -241,7 +131,7 @@ class FrameNetVisualizer:
         return result
     
     def create_dag_legend(self):
-        """Create legend elements for DAG visualization."""
+        """Create legend elements for FrameNet DAG visualization."""
         from matplotlib.patches import Patch
         return [
             Patch(facecolor='lightblue', label='Source Frames (no parents)'),
@@ -253,7 +143,7 @@ class FrameNetVisualizer:
         ]
     
     def create_taxonomic_legend(self):
-        """Create legend elements for taxonomic visualization."""
+        """Create legend elements for FrameNet taxonomic visualization."""
         from matplotlib.patches import Patch
         return [
             Patch(facecolor='lightblue', label='Root Frames (Depth 0)'),
@@ -261,185 +151,3 @@ class FrameNetVisualizer:
             Patch(facecolor='lightyellow', label='Level 2 Frames'),
             Patch(facecolor='lightcoral', label='Deeper Levels')
         ]
-    
-    def create_static_dag_visualization(self, save_path=None):
-        """Create a static DAG visualization using matplotlib."""
-        plt.figure(figsize=(16, 12))
-        
-        # Create DAG layout
-        pos = self.create_dag_layout()
-        
-        # Get node colors for DAG
-        node_colors = [self.get_dag_node_color(node) for node in self.G.nodes()]
-        
-        # Draw graph
-        nx.draw_networkx_nodes(self.G, pos, node_color=node_colors, node_size=2000, alpha=0.9)
-        nx.draw_networkx_labels(self.G, pos, font_size=8, font_weight='bold')
-        nx.draw_networkx_edges(self.G, pos, edge_color='gray', arrows=True, arrowsize=20, arrowstyle='->')
-        
-        plt.title(f"DAG {self.title}", fontsize=16, fontweight='bold')
-        plt.axis('off')
-        plt.tight_layout()
-        
-        # Add DAG legend
-        legend_elements = self.create_dag_legend()
-        plt.legend(handles=legend_elements, loc='upper right')
-        
-        # Save if path provided
-        if save_path:
-            plt.savefig(save_path, dpi=150, bbox_inches='tight')
-        
-        return plt
-    
-    def create_taxonomic_png(self, save_path):
-        """Generate a PNG for taxonomic (hierarchical) visualization."""
-        print(f"Generating taxonomic PNG visualization...")
-        
-        plt.figure(figsize=(16, 12))
-        
-        # Create taxonomic layout
-        pos = self.create_taxonomic_layout()
-        
-        # Get node colors for taxonomic visualization
-        node_colors = [self.get_taxonomic_node_color(node) for node in self.G.nodes()]
-        
-        # Draw hierarchical graph
-        nx.draw_networkx_nodes(self.G, pos, node_color=node_colors, node_size=2000, alpha=0.9)
-        nx.draw_networkx_labels(self.G, pos, font_size=8, font_weight='bold')
-        nx.draw_networkx_edges(self.G, pos, edge_color='gray', arrows=True, arrowsize=20, arrowstyle='->')
-        
-        plt.title(f"Taxonomic {self.title}", fontsize=16, fontweight='bold')
-        plt.axis('off')
-        plt.tight_layout()
-        
-        # Add taxonomic legend
-        legend_elements = self.create_taxonomic_legend()
-        plt.legend(handles=legend_elements, loc='upper right')
-        
-        # Save PNG
-        plt.savefig(save_path, dpi=150, bbox_inches='tight')
-        print(f"Saved taxonomic PNG to: {save_path}")
-        plt.close()
-    
-    def create_plotly_visualization(self, save_path=None, show=True):
-        """Create an interactive Plotly visualization."""
-        if not PLOTLY_AVAILABLE:
-            print("Warning: Plotly not available, falling back to static visualization")
-            return self.create_static_dag_visualization(save_path)
-        
-        # Create DAG layout
-        pos = self.create_dag_layout()
-        
-        # Prepare node data
-        node_x = []
-        node_y = []
-        node_text = []
-        node_color = []
-        hover_text = []
-        
-        for node in self.G.nodes():
-            x, y = pos[node]
-            node_x.append(x)
-            node_y.append(y)
-            node_text.append(node)
-            
-            # Color by DAG properties
-            node_color.append(self.get_dag_node_color(node))
-            
-            # Create hover text
-            if node in self.hierarchy:
-                data = self.hierarchy[node]
-                hover_info = [f"<b>{node}</b>"]
-                hover_info.append(f"Depth: {data.get('depth', 'Unknown')}")
-                
-                parents = data.get('parents', [])
-                if parents:
-                    hover_info.append(f"Parents: {', '.join(parents[:3])}")
-                
-                children = data.get('children', [])
-                if children:
-                    hover_info.append(f"Children: {', '.join(children[:5])}")
-                
-                # Add frame definition if available
-                frame_info = data.get('frame_info', {})
-                definition = frame_info.get('definition', '')
-                if definition and len(definition.strip()) > 0:
-                    if len(definition) > 150:
-                        definition = definition[:147] + "..."
-                    hover_info.append(f"<br>Definition: {definition}")
-            else:
-                hover_info = [f"<b>{node}</b>", "No additional information available."]
-            
-            hover_text.append('<br>'.join(hover_info))
-        
-        # Prepare edge data
-        edge_x = []
-        edge_y = []
-        
-        for edge in self.G.edges():
-            x0, y0 = pos[edge[0]]
-            x1, y1 = pos[edge[1]]
-            edge_x.extend([x0, x1, None])
-            edge_y.extend([y0, y1, None])
-        
-        # Create plotly figure
-        fig = go.Figure()
-        
-        # Add edges
-        fig.add_trace(go.Scatter(
-            x=edge_x, y=edge_y,
-            line=dict(width=2, color='gray'),
-            hoverinfo='none',
-            mode='lines',
-            name='Relations',
-            showlegend=False
-        ))
-        
-        # Add nodes
-        fig.add_trace(go.Scatter(
-            x=node_x, y=node_y,
-            mode='markers+text',
-            marker=dict(
-                size=20,
-                color=node_color,
-                line=dict(width=2, color='black')
-            ),
-            text=node_text,
-            textposition="middle center",
-            textfont=dict(size=10, color='black'),
-            hovertemplate='%{hovertext}<extra></extra>',
-            hovertext=hover_text,
-            name='Frames',
-            showlegend=False
-        ))
-        
-        # Update layout
-        fig.update_layout(
-            title=dict(text=f"DAG {self.title}", x=0.5, font=dict(size=16)),
-            showlegend=False,
-            hovermode='closest',
-            margin=dict(b=20,l=5,r=5,t=40),
-            annotations=[
-                dict(
-                    text="Hover over nodes for details | Zoom and pan to explore",
-                    showarrow=False,
-                    xref="paper", yref="paper",
-                    x=0.005, y=-0.002,
-                    xanchor='left', yanchor='bottom',
-                    font=dict(color='gray', size=10)
-                )
-            ],
-            xaxis=dict(showgrid=False, zeroline=False, showticklabels=False),
-            yaxis=dict(showgrid=False, zeroline=False, showticklabels=False),
-            plot_bgcolor='white'
-        )
-        
-        # Save HTML if path provided
-        if save_path:
-            fig.write_html(save_path)
-        
-        # Show if requested
-        if show:
-            fig.show()
-        
-        return fig
