+++ src/uvi/visualizations/__init__.py
@@ -5,8 +5,10 @@ This module provides classes for creating various visualizations of semantic gra
 including FrameNet and WordNet visualizations, DAG visualizations, taxonomic hierarchies, and interactive plots.
 """
 
+from .Visualizer import Visualizer
+from .InteractiveVisualizer import InteractiveVisualizer
 from .FrameNetVisualizer import FrameNetVisualizer
 from .InteractiveFrameNetGraph import InteractiveFrameNetGraph
 from .WordNetVisualizer import WordNetVisualizer
 
-__all__ = ['FrameNetVisualizer', 'InteractiveFrameNetGraph', 'WordNetVisualizer']
+__all__ = ['Visualizer', 'InteractiveVisualizer', 'FrameNetVisualizer', 'InteractiveFrameNetGraph', 'WordNetVisualizer']
